backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
bcrypt==3.2.2
email-validator
pandas
openpyxl
PyJWT
requests
pytest
httpx
